fix: harden log search inputs

+++ turbo/packages/api-contracts/src/contracts/runs.ts
@@ -956,9 +956,9 @@ export const logsSearchContract = c.router({
       agentId: z.string().uuid().optional(),
       runId: z.string().optional(),
       since: timestampQueryNumberSchema.optional(),
-      limit: z.coerce.number().min(1).max(50).default(20),
-      before: z.coerce.number().min(0).max(10).default(0),
-      after: z.coerce.number().min(0).max(10).default(0),
+      limit: z.coerce.number().int().min(1).max(50).default(20),
+      before: z.coerce.number().int().min(0).max(10).default(0),
+      after: z.coerce.number().int().min(0).max(10).default(0),
     }),
     responses: {
       200: logsSearchResponseSchema,
